+++ scripts/removeResourceFromWhitelist.ts
@@ -1,19 +1,28 @@
 import { ethers } from 'hardhat';
+import { IExecWhitelist } from '../typechain-types/contracts/IExecWhitelist.js';
 
 async function main() {
   const { CONTRACT_ADDRESS, ADDRESS_TO_REMOVE } = process.env;
+
+  if (!CONTRACT_ADDRESS) {
+    throw Error('missing env CONTRACT_ADDRESS');
+  }
+  if (!ADDRESS_TO_REMOVE) {
+    throw Error('missing env ADDRESS_TO_REMOVE');
+  }
 
   // Fetch the signer. Here, we're using the default signer (account[0]), but you can change this.
   const [signer] = await ethers.getSigners();
 
   // Get a contract instance
-  const IExecWhitelist = await ethers.getContractFactory('IExecWhitelist');
-  const whitelist = IExecWhitelist.attach(CONTRACT_ADDRESS as string).connect(
+  const IExecWhitelistFactory =
+    await ethers.getContractFactory('IExecWhitelist');
+  const whitelist = IExecWhitelistFactory.attach(CONTRACT_ADDRESS).connect(
     signer
-  );
+  ) as IExecWhitelist;
 
-  // Call the remRessourceToWhitelist function
-  const tx = await whitelist.remRessourceToWhitelist(ADDRESS_TO_REMOVE);
+  // Call the removeResourceFromWhitelist function
+  const tx = await whitelist.removeResourceFromWhitelist(ADDRESS_TO_REMOVE);
 
   // Wait for the transaction to be mined
   await tx.wait();