+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 # Changelog
 
+## [0.2.0]
+
+### Changed
+
+- **\[BREAKING\]** Fixed method name `addResourceToWhitelist`
+- **\[BREAKING\]** Fixed method name `removeResourceFromWhitelist`
+- Throw an error early when a script param is missing
+- Fixed TS typing errors
+
 ## [0.1.0] Initial release